pyproject.toml
[project]
name = "pokegrid-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Joshua Carney", email = "joshuakcarney@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "aiopokeapi>=0.1.12",
]

[project.scripts]
pokegrid-solver = "pokegrid_solver:main"

[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"
